주세은 : 보스스테이지 물풍선이랑 플레이어 물풍선 오류 업데이트

+++ CrazyArcade/GameEngineContents/BossBoom.cpp
@@ -44,18 +44,9 @@ void BossBoom::Update()
 							TileIndex _TileIndex = _Area.GetMapTile()->GetTileIndex(_Center);
 							BlockTile* _check = _Area.GetMapTile()->GetTile<BlockTile>(_TileIndex.X, _TileIndex.Y);
 							_check->TilePos_ = _Center;
-							if (_check->BlockType_ == BlockType::BubbleBlock || _check->BlockType_ == BlockType::WaveBlock)
-							{
-								BossBubblePop(_Center);
-								return;
-							}
-							
-							else
-							{
-								_check->Renderer->ChangeAnimation("Death");
-								_check->TileIndex_ = _TileIndex;
-								DeleteTileList_.push_back(_check);
-							}
+							_check->Renderer->ChangeAnimation("Death");
+							_check->TileIndex_ = _TileIndex;
+							DeleteTileList_.push_back(_check);
 						}
 
 
@@ -63,23 +54,29 @@ void BossBoom::Update()
 						{
 							GameEngineSound::SoundPlayOneShot("Boss_Hit.mp3");
 							TileIndex _TileIndex = _Area.GetMapTile()->GetTileIndex(_Center);
+							BlockTile* _check = _Area.GetMapTile()->GetTile<BlockTile>(_TileIndex.X, _TileIndex.Y);
+							
+
 							ItemBlockTile* _checkItem = _Area.GetMapTile()->GetTile<ItemBlockTile>(_TileIndex.X, _TileIndex.Y);
 							if (_checkItem != nullptr)
 							{
-								_Area.GetMapTile()->DeleteTile(_TileIndex.X, _TileIndex.Y);
+								if (_checkItem->BlockType_ == BlockType::ItemBlock)
+								{
+									_Area.GetMapTile()->DeleteTile(_TileIndex.X, _TileIndex.Y);
+									BossBubblePop(_Center);
+
+
+								}
+								
+							}
+							else
+							{
+								BossBubblePop(_Center);
 
 							}
-							BossBubblePop(_Center);
 						}
 
-						else if (3 == _Area.ChooseWaterAttackAni()) // 블럭 있음
-						{
-							GameEngineSound::SoundPlayOneShot("Boss_Hit.mp3");
-							TileIndex _TileIndex = _Area.GetMapTile()->GetTileIndex(_Center);
-							ItemBlockTile* _checkItem = _Area.GetMapTile()->GetTile<ItemBlockTile>(_TileIndex.X, _TileIndex.Y);
-							_Area.GetMapTile()->DeleteTile(_TileIndex.X, _TileIndex.Y);
-							BossBubblePop(_Center);
-						}
+
 					}
 				}
 				SetPosition(float4::ZERO);
